Add toSmithyTraits and toSmithy4sHints to internals.conversion.syntax
