+++ modules/dynamic/src-jvm/DynamicSchemaIndexPlatform.scala
@@ -24,7 +24,7 @@ import software.amazon.smithy.model.shapes.OperationShape
 import scala.jdk.CollectionConverters._
 import software.amazon.smithy.model.shapes.ServiceShape
 import software.amazon.smithy.model.Model
-import smithy4s.dynamic.syntax._
+import smithy4s.dynamic.internals.conversion.syntax._
 
 private[dynamic] trait DynamicSchemaIndexPlatform {
   self: DynamicSchemaIndex =>
